+++ tests/Imagine/Filter/PostProcessor/AbstractPostProcessorTestCase.php
@@ -60,10 +60,6 @@ abstract class AbstractPostProcessorTestCase extends AbstractTest
         }
 
         $m = $r->getMethod($method);
-        // remove when we drop support for PHP older than 8.1
-        if (PHP_VERSION_ID < 80100) {
-            $m->setAccessible(true);
-        }
 
         return $m;
     }
@@ -80,10 +76,6 @@ abstract class AbstractPostProcessorTestCase extends AbstractTest
         }
 
         $p = $r->getProperty($property);
-        // remove when we drop support for PHP older than 8.1
-        if (PHP_VERSION_ID < 80100) {
-            $p->setAccessible(true);
-        }
 
         return $p;
     }